Show error when configured hotkey is already taken.
